Create CDF-to-HDF5.md
Created a draft of CDF to HDF5 mapping. This document should be used as a reference for the CDF VOL connector. We will not need to implement the full mapping for reading data with a VOL connector, but we will need it for the converter in the future.

We will create another document that will reflect CDF reader implementation via HDF5 VOL.